*** empty log message ***

+++ java/tests_core/net/jsunit/model/BrowserResultTest.java
@@ -1,14 +1,10 @@
 package net.jsunit.model;
 
-import junit.framework.TestCase;
-
 import java.io.File;
 import java.util.List;
 
+import junit.framework.TestCase;
 import net.jsunit.Utility;
-import net.jsunit.model.BrowserResult;
-import net.jsunit.model.TestCaseResult;
-import org.jdom.Element;
 
 /**
  * @author Edward Hieatt, edward@jsunit.net
@@ -16,22 +12,26 @@ import org.jdom.Element;
 
 public class BrowserResultTest extends TestCase {
     private BrowserResult result;
-    private String expectedXmlFragment =
-            "<testrun id=\"An ID\" errors=\"1\" failures=\"1\" name=\"JsUnitTestCase\" tests=\"3\" time=\"4.3\">"
-                + "<properties>"
-                    + "<property name=\"JsUnitVersion\" value=\"2.5\" />"
-                    + "<property name=\"userAgent\" value=\"Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1)\" />"
-                    + "<property name=\"remoteAddress\" value=\"Dummy Remote Address\" />"
-                    + "<property name=\"baseURL\" value=\"http://www.example.com/\" />"
-                + "</properties>"
-                    + "<testcase name=\"page1.html:testFoo\" time=\"1.3\" />"
-                    + "<testcase name=\"page1.html:testFoo\" time=\"1.3\">"
-                        + "<error>Test Error Message</error>"
-                    + "</testcase>"
-                    + "<testcase name=\"page2.html:testFoo\" time=\"1.3\">"
-                        + "<failure>Test Failure Message</failure>"
-                    + "</testcase>"
-            + "</testrun>";
+    
+    private static String expectedXmlFragment =
+        "<browserResult id=\"An ID\" time=\"4.3\">" +
+        	"<properties>" +
+            	"<property name=\"browserFileName\" value=\"c:\\Program Files\\Internet Explorer\\iexplore.exe\" />" +
+            	"<property name=\"jsUnitVersion\" value=\"2.5\" />" +
+                "<property name=\"userAgent\" value=\"Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1)\" />" +
+                "<property name=\"remoteAddress\" value=\"Dummy Remote Address\" />" +
+                "<property name=\"baseURL\" value=\"http://www.example.com/\" />" +
+            "</properties>" +
+            "<testCases>" +
+	            "<testCase name=\"page1.html:testFoo\" time=\"1.3\" />" +
+	            "<testCase name=\"page1.html:testFoo\" time=\"1.3\">" +
+	                "<error>Test Error Message</error>" +
+	            "</testCase>" +
+	            "<testCase name=\"page2.html:testFoo\" time=\"1.3\">" +
+	                "<failure>Test Failure Message</failure>" +
+	            "</testCase>" +
+            "</testCases>" +
+        "</browserResult>";
 
     public void setUp() throws Exception {
         super.setUp();
@@ -42,7 +42,7 @@ public class BrowserResultTest extends TestCase {
             "page2.html:testFoo|1.3|F|Test Failure Message|"}
         );
     }
-
+    
     public void testId() {
         assertNotNull(result.getId());
         result = new BrowserResult();
@@ -68,13 +68,22 @@ public class BrowserResultTest extends TestCase {
     }
 
     public void testBuildFromXmlFile() {
-        Element element = result.asXml();
-        String xmlString = Utility.asString(element);
-        Utility.writeFile(xmlString, "resultXml.xml");
-        File file = new File("resultXml.xml");
-        BrowserResult reconstitutedResult = BrowserResult.fromXmlFile(file);
+    	File file = null;
+    	try {
+	        Utility.writeFile(expectedXmlFragment, "resultXml.xml");
+	        file = new File("resultXml.xml");
+	        BrowserResult reconstitutedResult = new BrowserResultBuilder().build(file);
+	        assertEquals(BrowserResult.class, reconstitutedResult.getClass());
+	        assertFields(reconstitutedResult);
+    	} finally {
+    		if (file != null)
+    			file.delete();
+    	}
+    }
+
+    public void testBuildFromXmlDocument() {
+        BrowserResult reconstitutedResult = new BrowserResultBuilder().build(result.asXmlDocument());
         assertFields(reconstitutedResult);
-        file.delete();
     }
 
     public void testFailure() {
@@ -107,9 +116,16 @@ public class BrowserResultTest extends TestCase {
     	assertEquals("page2.html", result2.getTestPageName());
     	assertEquals(1, result2.getTestCaseResults().size());
     }
-
+    
+    public void testCompleted() {
+    	assertTrue(result.completedTestRun());
+    	assertFalse(result.timedOut());
+    	assertFalse(result.failedToLaunch());
+    }
+    
     private BrowserResult createBrowserResult() {
         BrowserResult browserResult = new BrowserResult();
+        browserResult.setBrowserFileName("c:\\Program Files\\Internet Explorer\\iexplore.exe");
         browserResult.setJsUnitVersion("2.5");
         browserResult.setId("An ID");
         browserResult.setUserAgent("Mozilla/4.0 (compatible; MSIE 6.0; Windows NT 5.1)");
